Create core.md
A very early and rough API document for the core functionality in Extempore (e.g. the standard library).

This is not ready for prime time, but shows how the API documentation going forward could work. There's certain problems that need to be solved. In particular:
+ Where do Scheme and Extempore API documents go? Same API documents, or different ones?
+ How to handle polymorphic functions and macros

Consider this a prototype for something we can (at least partially) automate going forward.

Currently string and formatting documentation is missing.