Create audio_dsp.md
This is a early version of an API document for audio_dsp. Quite a bit is currently missing, and certain parts are in (friendly) disagreement with certain decisions made with this API ;->

Nonetheless I'd like to discuss some of it, as I think it's a useful first step to defining this stuff. Issues that emerged include:
+ How to document closure properties (and what should be documented)
+ Which functions should be ignored by documentation (e.g. helper functions that are not part of the 'external interface')