Rough draft on memory management
I'm parking this here for the moment. I now think this is the wrong way to discuss memory management, but I don't want to lose what I've written to date.

Memory management with raw data is going to come a LOT later in the documentation. But stack, zones and heap needs to come after types.