remove flanger from making an instrument guide
for now - flanger_c changed a while ago (with another bunch of instrument
changes) and now it's a bit richer, but perhaps a bit too complicated for a
"getting started" guide like this?

Anyway - it's removed for now because the current version was broken - we could
always fix in the future.

closes #23